requirements.txt
# Requirements

# Dependencis
composio-openai==0.11.1
openai
python-dotenv

# Requirements
Requirement already satisfied: composio-openai==0.11.1 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from -r requirements.txt (line 4)) (0.11.1)
Requirement already satisfied: openai in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from -r requirements.txt (line 5)) (2.24.0)
Requirement already satisfied: python-dotenv in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from -r requirements.txt (line 6)) (1.2.1)
Requirement already satisfied: composio in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from composio-openai==0.11.1->-r requirements.txt (line 4)) (0.11.1)
Requirement already satisfied: anyio<5,>=3.5.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (4.12.1)
Requirement already satisfied: distro<2,>=1.7.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (1.9.0)
Requirement already satisfied: httpx<1,>=0.23.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (0.28.1)
Requirement already satisfied: jiter<1,>=0.10.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (0.13.0)
Requirement already satisfied: pydantic<3,>=1.9.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (2.12.5)
Requirement already satisfied: sniffio in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (1.3.1)
Requirement already satisfied: tqdm>4 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (4.67.3)
Requirement already satisfied: typing-extensions<5,>=4.11 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from openai->-r requirements.txt (line 5)) (4.15.0)
Requirement already satisfied: idna>=2.8 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from anyio<5,>=3.5.0->openai->-r requirements.txt (line 5)) (3.11)
Requirement already satisfied: certifi in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from httpx<1,>=0.23.0->openai->-r requirements.txt (line 5)) (2026.2.25)
Requirement already satisfied: httpcore==1.* in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from httpx<1,>=0.23.0->openai->-r requirements.txt (line 5)) (1.0.9)
Requirement already satisfied: h11>=0.16 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from httpcore==1.*->httpx<1,>=0.23.0->openai->-r requirements.txt (line 5)) (0.16.0)
Requirement already satisfied: annotated-types>=0.6.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from pydantic<3,>=1.9.0->openai->-r requirements.txt (line 5)) (0.7.0)
Requirement already satisfied: pydantic-core==2.41.5 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from pydantic<3,>=1.9.0->openai->-r requirements.txt (line 5)) (2.41.5)
Requirement already satisfied: typing-inspection>=0.4.2 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from pydantic<3,>=1.9.0->openai->-r requirements.txt (line 5)) (0.4.2)
Requirement already satisfied: pysher>=1.0.8 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (1.0.8)
Requirement already satisfied: composio-client==1.27.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (1.27.0)
Requirement already satisfied: json-schema-to-pydantic>=0.4.8 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (0.4.9)
Requirement already satisfied: websocket-client!=0.49 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from pysher>=1.0.8->composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (1.9.0)
Requirement already satisfied: requests>=2.26.0 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from pysher>=1.0.8->composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (2.32.5)
Requirement already satisfied: charset_normalizer<4,>=2 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from requests>=2.26.0->pysher>=1.0.8->composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (3.4.4)
Requirement already satisfied: urllib3<3,>=1.21.1 in /Users/rnemzek/Library/Python/3.12/lib/python/site-packages (from requests>=2.26.0->pysher>=1.0.8->composio->composio-openai==0.11.1->-r requirements.txt (line 4)) (2.6.3)

